requirements.txt
defusedxml==0.7.1
fpdf2==2.5.1
Pillow==9.1.0
PyPDF2==1.27.7
